Organize and document Scala 3 derivation utils
