+++ core/src/main/scala-3/pureconfig/generic/derivation/ConfigReaderDerivation.scala
@@ -2,6 +2,7 @@ package pureconfig
 package generic
 package derivation
 
+import scala.compiletime.summonFrom
 import scala.deriving.Mirror
 
 trait ConfigReaderDerivation extends CoproductConfigReaderDerivation with ProductConfigReaderDerivation {
@@ -11,6 +12,15 @@ trait ConfigReaderDerivation extends CoproductConfigReaderDerivation with Produc
         case given Mirror.ProductOf[A] => derivedProduct
         case given Mirror.SumOf[A] => derivedSum
       }
+  }
+
+  /** Summons a `ConfigReader` for a given type `A`. It first tries to find an existing given instance of
+    * `ConfigReader[A]`. If none is found, it tries to derive one using this `ConfigReaderDerivation` instance. This
+    * method differs from `derived` in that the latter doesn't try to find an existing instance first.
+    */
+  protected inline def summonConfigReader[A] = summonFrom {
+    case reader: ConfigReader[A] => reader
+    case given Mirror.Of[A] => ConfigReader.derived[A]
   }
 }
