--- core/src/main/scala-3/pureconfig/generic/derivation/Labels.scala
@@ -1,17 +1,0 @@
-package pureconfig.generic
-package derivation
-
-import scala.compiletime.{constValue, erasedValue, summonFrom, summonInline}
-import scala.deriving.Mirror
-
-object Labels {
-  inline def transformed[T <: Tuple](
-      inline transform: String => String
-  ): List[String] =
-    inline erasedValue[T] match {
-      case _: (h *: t) =>
-        transform(constValue[h & String]) :: transformed[t](transform)
-
-      case _: EmptyTuple => Nil
-    }
-}
