--- core/src/main/scala-3/pureconfig/generic/derivation/WidenType.scala
@@ -1,8 +1,0 @@
-package pureconfig.generic.derivation
-
-object WidenType {
-  inline def widen[A, B](a: A): A & B =
-    inline a match {
-      case b: B => b
-    }
-}
